pyproject.toml
[tool.poetry]
name = "sttntts"
version = "0.1.0"
description = ""
authors = ["Chaiwat Saehui <art_of_destroy@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
speechrecognition = "^3.10.4"
pyaudio = "0.2.14"
google-cloud-speech = "^2.26.0"
openai = "^1.34.0"
install = "^1.3.5"
google-cloud-texttospeech = "^2.16.3"
twitchapi = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
